Add RDDRONE-T1ADAPT v1.0 release
RDDRONE-T1ADAPT Ethernet adapter board is evaluation platform for T1
Ethernet. This network media converter allows 100BASE-T1 automotive
devices and devices with traditional 100BASE-TX incorporating an RJ45
connector to be combined in a single network.
This v1.0 release of the adapter board supports 100 Mbit/s links on
RJ45 and T1 connectors. Automatic mode detection is enabled on the T1
connectors, so no further adjustments are needed.

Signed-off-by: Anjali Namdev <anjali.namdev@nxp.com>